SellFruit
과일 판매창&판매 기능 구현
SlotManager 클래스 만들어서 다형성 이용했다.

기존의 BuySeedSlotManager와 SellFruitSlotManager 가 SlotManager 클래스를 상속받도록 했다.

+++ Assets/02.Scripts/Farm/Fruit.cs
@@ -8,7 +8,7 @@ public class Fruit : MonoBehaviour
     // 만들어놓은 프리팹은 FruitContainer 에 저장할 것..
 
     public string fruitName;
-    public float fruitPrice;
+    public int fruitPrice;
     public bool isEnabled = false;
 
     public int fruitIdx = 0; // 과일 인덱스
